Move the browser tests into a subfoler.
