+++ package.json
@@ -27,9 +27,9 @@
     "pretest": "npm run --silent lint",
     "preversion": "npm run check-changelog && npm run check-only-changelog-changed",
     "tag": "git tag v$npm_package_version",
-    "test": "npm run testonly",
-    "testonly": "karma start --single-run",
-    "testonly:watch": "karma start --no-single-run",
+    "test": "npm run test:browser",
+    "test:browser": "karma start --single-run",
+    "test:browser:watch": "karma start --no-single-run",
     "performance-test:watch": "webpack --watch --config webpack.config.performance-test.js"
   },
   "author": "Brigade Engineering",
