+++ tests.webpack.js
@@ -1,2 +1,2 @@
-const context = require.context('./test', true, /_test\.jsx?$/);
+const context = require.context('./test/browser', true, /_test\.jsx?$/);
 context.keys().forEach(context);
